### 2025-11-17 #9437 @editorialbot
[PRE REVIEW]: `FastPCA`: An R package for fast singular value decomposition
**Submitting author:** @ACSoupir (<a href="http://orcid.org/0000-0003-1251-9179">Alex Soupir</a>)
**Repository:** https://github.com/ACSoupir/FastPCA
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.2
