+++ paper/paper.md
@@ -37,7 +37,7 @@ affiliations:
   - name: Department of Bioinformatics and Biostatistics, Moffitt Cancer Center
     index: 5
 citation_author: Ward et al
-date: "03 November, 2025"
+date: "17 November, 2025"
 year: "2025"
 bibliography: paper.bib
 output: 
