+++ paper/paper.bib
@@ -45,6 +45,7 @@
   pages    = "308",
   month    =  dec,
   year     =  2024,
+  doi = {10.1186/s13059-024-03435-z},
   keywords = "Immunotherapy resistance; Ligand receptor; Malignant-cell typing;
               Single-cell RNA; Spatial transcriptomics",
   language = "en"
@@ -73,6 +74,7 @@
   pages    = "47",
   month    =  jan,
   year     =  2025,
+  doi = {10.1186/s12864-025-11235-3},
   keywords = "10X Visium; 10X visium HD; CosMx SMI; GeoMx DSP; Merscope; RNA
               sequencing; Spatial transcriptomics; Stereoseq; Xenium",
   language = "en"
@@ -104,6 +106,7 @@
   pages     = "5040--5055",
   month     =  dec,
   year      =  2010,
+  doi = {10.3390/ijms11125040},
   keywords  = "biomarker; imaging mass spectrometry; lipids; matrix-assisted
                laser desorption/ionization mass spectrometry; protein",
   copyright = "https://creativecommons.org/licenses/by/3.0/",
@@ -117,6 +120,7 @@
   publisher = "The R Foundation",
   month     =  may,
   year      =  2011,
+  doi = {10.32614/CRAN.package.irlba},
   note      = "Title of the publication associated with this dataset: CRAN:
                Contributed Packages"
 }
@@ -157,6 +161,7 @@
   pages    = "1599--1608",
   month    =  sep,
   year     =  2023,
+  doi = {10.1101/gr.277525.122},
   language = "en"
 }
 
@@ -177,6 +182,7 @@
                   for a standard decomposition of spectral data cubes. The
                   qrpca source code is made freely available to the community.",
   year         =  2022,
+  doi = {arXiv:2206.06797},
   primaryClass = "astro-ph.IM",
   eprint       = "2206.06797"
 }
@@ -187,6 +193,7 @@
   publisher = "The R Foundation",
   month     =  aug,
   year      =  2020,
+  doi = {10.32614/CRAN.package.torch},
   note      = "Title of the publication associated with this dataset: CRAN:
                Contributed Packages"
 }
@@ -219,6 +226,7 @@
                   of individual subsystems, as well as the overall speed of
                   PyTorch on several common benchmarks.",
   year         =  2019,
+  doi = {arXiv:1912.01703},
   primaryClass = "cs.LG",
   eprint       = "1912.01703"
 }
@@ -229,6 +237,7 @@
   publisher = "The R Foundation",
   month     =  mar,
   year      =  2017,
+  doi = {10.32614/CRAN.package.reticulate},
   note      = "Title of the publication associated with this dataset: CRAN:
                Contributed Packages"
 }
