+++ paper/paper.bib
@@ -166,25 +166,18 @@
 }
 
 @ARTICLE{De_Souza2022-lk,
-  title        = "Qrpca: A package for fast principal component analysis with
-                  {GPU} acceleration",
-  author       = "de Souza, Rafael S and Quanfeng, Xu and Shen, Shiyin and
-                  Peng, Chen and Mu, Zihao",
-  abstract     = "We present qrpca, a fast and scalable QR-decomposition
-                  principal component analysis package. The software, written
-                  in both R and python languages, makes use of torch for
-                  internal matrix computations, and enables GPU acceleration,
-                  when available. qrpca provides similar functionalities to
-                  prcomp (R) and sklearn (python) packages respectively. A
-                  benchmark test shows that qrpca can achieve computational
-                  speeds 10-20 $\times$ faster for large dimensional matrices
-                  than default implementations, and is at least twice as fast
-                  for a standard decomposition of spectral data cubes. The
-                  qrpca source code is made freely available to the community.",
-  year         =  2022,
-  doi = {arXiv:2206.06797},
-  primaryClass = "astro-ph.IM",
-  eprint       = "2206.06797"
+title = {qrpca: A package for fast principal component analysis with GPU acceleration},
+journal = {Astronomy and Computing},
+volume = {41},
+pages = {100633},
+year = {2022},
+issn = {2213-1337},
+doi = {10.1016/j.ascom.2022.100633},
+url = {https://www.sciencedirect.com/science/article/pii/S221313372200052X},
+author = {R. {S. de Souza} and X. Quanfeng and S. Shen and C. Peng and Z. Mu},
+keywords = {Principal component analysis, Astroinformatics, GPU computing},
+abstract = {We present qrpca, a fast and scalable QR-decomposition principal component analysis package. The software, written in both R and python languages, makes use of torch for internal matrix computations, and enables GPU acceleration, when available. qrpca provides similar functionalities to prcomp (R) and sklearn (python) packages respectively. A benchmark test shows that qrpca can achieve computational speeds 10–20 × faster for large dimensional matrices than default implementations, and is at least twice as fast for a standard decomposition of spectral data cubes. The qrpca source code is made freely available to the community.}
+
 }
 
 @MISC{Falbel2020-wq,
@@ -226,7 +219,7 @@
                   of individual subsystems, as well as the overall speed of
                   PyTorch on several common benchmarks.",
   year         =  2019,
-  doi = {arXiv:1912.01703},
+  doi = {10.48550/arXiv.1912.01703},
   primaryClass = "cs.LG",
   eprint       = "1912.01703"
 }
