### 2025-12-23 #9437 @staradutt
Hi @rich2355 , I haven't ever worked with R and hence would like to skip. Thank you !
