### 2025-12-23 #9437 @rich2355
Hi @matt-graham, thanks for letting me know!

Hi @dhjx1996, thanks for accepting my invitation! The review process will actually begin when I confirm all the reviewers. I will just wait for some time (maybe 3 days?) to see whether there will be a third reviewer coming up. In the meantime, you can definitely start reviewing and taking notes, and you can communicate these with the authors when the review starts. Thanks again!
