### 2026-01-07 #9706 @dhjx1996
> Hi [@dhjx1996](https://github.com/dhjx1996), thanks for asking! Yes, the new requirements became effective on Jan 6. Therefore, please follow the new ones.

Alright. In that case, @ACSoupir from my initial look at your paper it seems like you will have to make some changes to comply with the new requirements.

In addition, @rich2355 may I clarify for @ACSoupir how may he satisfy the "Open development" and "Collaborative effort" requirements? The repo currently shows 5 months of commits and they are all from @ACSoupir, though the paper does have 6 authors.
