### 2026-01-07 #9706 @dhjx1996
> > Thanks [@rich2355](https://github.com/rich2355) and [@arfon](https://github.com/arfon)! Is there a way for me to generate the old checklist?
>
> It seemed to be updated?

Ah, apologies, I did not realize. I'll proceed with the review then.
