### 2026-01-11 #9706 @chrisarg
Thank you for sending this over for review. I really enjoyed it and I can see the application potential in the field of bioinformatics and biomarker discovery.  However the "package" of repo+paper need considerable work to ensure reproducibility of analyses

A few comments for your consideration (I will intermingle package and repository comments), but since the material is reported in both places I will separate suggestions for one or both.

1. _Style_ : Can you ensure that color blind visual palette (such as _Viridis_, or _cividis_ is used throughout the **paper**)?
2.  One of the comparators is listed as  `BioSingular::runSVD` in some parts of the text, but shouldn't it be `BiocSingular::runSVD` throughout the **manuscript text**?
3. I noticed that the software used bench::mark() to carry out the timing and memory profiling. However this function only measures memory allocations through the garbage collector of R and not through external libraries that the R code interfaces with (in this case it appears that `libtorch` in Python is used, so that the memory usage is considerably under-estimated relative to packages that allocate their memory in a way that is managed by R). Furthermore, this memory counts allocations, not _memory use_ as suggested in line 89. To see the difference, consider the case of an iterative algorithm that uses 100 MB per iteration and (for whatever reason) frees and re-allocates the space per iteration. If the algorithm takes 100 iterations to converge, then bench::mark would report a 10GB memory allocation, but the algorithm only really used 100MB.
4. Have the authors profiled the RSS (Resident Set Size) to map peak memory usage through the OS? Since some libraries allocate through R and others through the library backends the memory comparisons may be comparing apples to oranges. Either remove the memory comparisons if you cannot guarantee the veracity of the memory benchmarks, or instrument the R benchmark code to use an OS level profiler to report peak memory use in addition to the allocations. In the latter case, feel free to reuse/draw inspiration from the instrumentation/profiler for R I wrote a few months ago [here](https://chrisarg.github.io/Killing-It-with-PERL/2025/01/18/Timing-Peak-DRAM-Use-In-R-With-Perl-Part-1.html) and [there](https://chrisarg.github.io/Killing-It-with-PERL/2025/01/19/Timing-Peak-DRAM-Use-In-R-With-Perl-Part-2.html)
5. There is s a short of a linear (in log space) , but very noisy relation between memory allocations and execution time suggested in Figure 1, and libraries that don't allocate through R (or at least don't trigger the wrath of the garbage collector!) achieve better performances. This may be worth commenting in the **manuscript text** if the memory profiles are kept (and it may be a good idea to include in the markdown of the repository).
6. I get a 404 error when I try to assess the [tutorial/vignette](https://github.com/ACSoupir/FastPCA/blob/main/articles/intro.html) mentionted at the repository website. Similarly, following the [link to the benchmarking script ](https://github.com/ACSoupir/FastPCA/docs_acs/paper/benchmarking_script.R) from within the PDF landed me to a **Not Found**
7. As the back-end is written in another language (not R), the description of the library in lines 102 is somewhat misleading: the API is R-native but the implementation is not. The library does provide a substantial contribution to PCA analyses for large datasets _using R_, but the compute heavy machinery is done in C++ (libtortch and torch). I think the words "R native" should be removed.
8. Some packages are missing citations e.g. `BiocSingular::runSVD`, `bigstatsr`, `pcaone` . Can you ensure that all comparators are cited appropriately?
9. Consider renaming "Demonstration of Improvement" to "Research Impact"
10. I could not execute the main vignette (`SpatialTranscriptomics.Rmd`)  in a clean environment with the instructions given. In particular, sourcing the Rds file generates the following error :
```r
Loading required namespace: SeuratObject
Failed with error:  ‘there is no package called ‘SeuratObject’’
```
because the package Seurat is not listed as a dependency for the vignette.  The same applies to the `intro.Rmd` vignette, which depends on `dplyr` and `ggplot2` (but neither is  mentioned as a dependency).
12. It may be worthwhile to provide a bash script to automate the analysis of the Spatial Transcriptomics vignetteby downloading from Zenodo, sourcing the R file and producing the results. Not absolutely necessary, but considering your main vignette lands into a 404, and that there are undocumented dependencies for at least one vignette in your repo, having such a script could be valuable.
