### 2026-01-23 12cefff
add checks for suggested packages (only needed for the vignette) and echoed the example matrix creation code.
