### 2026-01-23 28ef2d5
move umap-learn to setup when not needing to create a new environment.
