### 2026-01-26 6290aac
add GPU results to benchmarks using kidney data
