+++ paper/paper.md
@@ -37,8 +37,8 @@ affiliations:
   - name: Department of Bioinformatics and Biostatistics, Moffitt Cancer Center
     index: 5
 citation_author: Ward et al
-date: "17 November, 2025"
-year: "2025"
+date: "26 January, 2026"
+year: "2026"
 bibliography: paper.bib
 output: 
   rticles::joss_article:
@@ -71,63 +71,83 @@ Then, there are processing functions that work on input matrices or intermediate
 -   `get_pc_scores()`: calculates the principal component scores from the output of `FastPCA()` ($U$ matrix which contains left singular vectors, $S$ vector containing the singular values, and $V^T$ matrix of the right singular vectors) which are typically expected for downstream analyses
 -   `umap()`: uses either `uwot` (R) or `umap-learn` (python) for visualization of the principal component scores
 
-# Demonstration of Improvement
+# Research Impact
 
-To demonstrate the improvements of `FastPCA`, we use our previous data set of single-cell spatial transcriptomics of kidney cancer. The dataset is publicly available on [Zenodo](https://zenodo.org/doi/10.5281/zenodo.12730226) and can be downloaded locally to be used (@Soupir2024-ww). Since the main benefits of `FastPCA` are in its application of singular value decomposition, that is what will be focused on. The counts for each cell were extracted from the Seurat object in the Nanostring assay, which contains the expression of 978 probes (959 genes, 19 negative control probes) in 199,112 cells. Normalization was performed using `prep_matrix()`, applying a log transformation, scaling (mean centering and unit variance), and transposing for samples to be rows and columns to be the gene features. The full script can be found <https://github.com/ACSoupir/FastPCA/docs_acs/paper/benchmarking_script.R>.
+To demonstrate the improvements of `FastPCA`, we use our previous data set of single-cell spatial transcriptomics of kidney cancer. The dataset is publicly available on [Zenodo](https://zenodo.org/doi/10.5281/zenodo.12730226) and can be downloaded locally to be used (@Soupir2024-ww). Since the main benefits of `FastPCA` are in its application of singular value decomposition, that is what will be focused on. The counts for each cell were extracted from the Seurat object in the Nanostring assay, which contains the expression of 978 probes (959 genes, 19 negative control probes) in 199,112 cells. Normalization was performed using `prep_matrix()`, applying a log transformation, scaling (mean centering and unit variance), and transposing for samples to be rows and columns to be the gene features. The full script can be found <https://github.com/ACSoupir/FastPCA/blob/main/docs_acs/paper/benchmarking_script.R>.
 
-Experiments were performed using `FastPCA`, both randomized SVD and exact, `irlba`, `pcaone`, and `bigstatsr`, both partial and randomized partial SVD. Performance measures were collected from a M3 Pro MacBook Pro with 36GB unified RAM. `FastPCA` (4 cores), `bigstatsr` partial (1 core) and random (4 cores), `irlba` (1 core), and `pcaone` (1 core) were assessed. Time was profiled with `bench::mark()` function. The elapsed time of `FastPCA` calculating the singular value decomposition on the full data took 10.74 seconds on average and the randomized SVD with `FastPCA` taking 8.46 seconds on average (**Table 1**, **Figure 1**). The next fastest were the methods with PCAone which took 27.29 and 29.58 seconds on average for "alg1" and "alg2", respectively.
[…]
