+++ paper/paper.bib
@@ -247,3 +247,100 @@ abstract = {We present qrpca, a fast and scalable QR-decomposition principal com
   URL = {https://doi.org/10.1137/090771806},
   eprint = {https://doi.org/10.1137/090771806}
 }
+
+@Manual{biocsingular,
+  title = {BiocSingular: Singular Value Decomposition for Bioconductor Packages},
+  author = {Aaron Lun},
+  year = {2024},
+  note = {R package version 1.22.0},
+  url = {https://bioconductor.org/packages/BiocSingular},
+  doi = {10.18129/B9.bioc.BiocSingular},
+}
+
+@Article{prive2018efficient,
+  title = {Efficient analysis of large-scale genome-wide data with two {R} packages: bigstatsr and bigsnpr},
+  author = {Florian Priv{\'e} and Hugues Aschard and Andrey Ziyatdinov and Michael G.B. Blum},
+  journal = {Bioinformatics},
+  volume = {34},
+  number = {16},
+  pages = {2781--2787},
+  year = {2018},
+  publisher = {Oxford University Press},
+  doi = {10.1093/bioinformatics/bty185},
+  url = {https://doi.org/10.1093/bioinformatics/bty185},
+}
+
+@Manual{memprof,
+    title = {memprof: Profile memory over time of system or a process},
+    author = {Robert Ashton},
+    year = {2026},
+    note = {R package version 0.0.1, commit 1777ddfb3adbe408538f8fa424c9959bf39ea0ad},
+    url = {https://github.com/mrc-ide/memprof},
+  }
+
+@ARTICLE{Eckart1936-mg,
+  title     = "The approximation of one matrix by another of lower rank",
+  author    = "Eckart, Carl and Young, Gale",
+  abstract  = "The mathematical problem of approximating one matrix by another
+               of lower rank is closely related to the fundamental postulate of
+               factor-theory. When formulated as a least-squares problem, the
+               normal equations cannot be immediately written down, since the
+               elements of the approximate matrix are not independent of one
+               another. The solution of the problem is simplified by first
+               expressing the matrices in a canonic form. It is found that the
+               problem always has a solution which is usually unique. Several
+               conclusions can be drawn from the form of this solution.A
+               hypothetical interpretation of the canonic components of a score
+               matrix is discussed.",
+  journal   = "Psychometrika",
+  publisher = "Springer Science and Business Media LLC",
+  volume    =  1,
+  number    =  3,
+  pages     = "211--218",
+  month     =  sep,
+  year      =  1936,
+  language  = "en"
+}
+
+@Manual{desctools,
+    title = {DescTools: Tools for Descriptive Statistics},
+    author = {Andri Signorell},
+    year = {2025},
+    note = {R package version 0.99.60},
+    url = {https://CRAN.R-project.org/package=DescTools},
+  }
+
+
+@ARTICLE{Akoglu2018-pq,
+  title     = "User's guide to correlation coefficients",
+  author    = "Akoglu, Haldun",
+  abstract  = "When writing a manuscript, we often use words such as perfect,
+               strong, good or weak to name the strength of the relationship
+               between variables. However, it is unclear where a good
+               relationship turns into a strong one. The same strength of r is
+               named differently by several researchers. Therefore, there is an
+               absolute necessity to explicitly report the strength and
+               direction of r while reporting correlation coefficients in
+               manuscripts. This article aims to familiarize medical readers
+               with several different correlation coefficients reported in
+               medical manuscripts, clarify confounding aspects and summarize
+               the naming practices for the strength of correlation
+               coefficients.",
[…]
