### 2026-01-26 b8fd679
make it so if calling a python backend, it doesn't trigger reticulate to load an environment for prep_matrix
