### 2026-01-26 c6c5aea
add check for data frame prefer conda for torch force initialization of python when starting cleaner warning start warning about pytorch or torch
