### 2026-01-26 67116b0
if torch after pytorch, send error. libomp.dylib makes mutually exclusive unless setting KMP_DUPLICATE_LIB_OK but potentially has unstable results
