### 2026-01-26 f2c19bc
update readme with dependencies/suggest and remove memory benchmarking
