### 2026-01-26 #9706 @ACSoupir
Thanks @rich2355. Am trying to finish up response to @dhjx1996 and @chrisarg. Agree that there were good comments.!
