### 2026-01-26 3fc5c28
code blocks for 'suggests' packages and updated github io articles
