### 2026-01-26 96117ba
default require initialization when running so that python is automatically loaded
