### 2026-01-26 111af48
fix column names from PC in the SVD to dim, used dim to PC in get_pc_scores
