### 2026-01-26 bf277db
add tests per reviewer suggestions
