+++ paper/ReviewerComments1.Rmd
@@ -1,0 +1,98 @@
+---
+output: html
+---
+
+# @chrisarg
+
+> Point 1: Color Blind Palette for plots (viridis)
+
+Thank you for the suggestion. Have updated Figure 1 to use the Viridis color palette as suggested. 
+
+> Point 2: Correct spelling of `BioSingular::runSVD` to `BiocSingular::runSVD`
+
+We have corrected the in text spelling of the package. Thank you for pointing this out.
+
+> Points 3 and 4: Profile memory using @chrisarg profiler [here](https://chrisarg.github.io/Killing-It-with-PERL/2025/01/18/Timing-Peak-DRAM-Use-In-R-With-Perl-Part-1.html)
+
+Have attempted to profile memory using both your perl method as well as an R package (`memprof`) that claims to track the main process as well as any child processes (which by looking at the code in your [part 2](https://chrisarg.github.io/Killing-It-with-PERL/2025/01/19/Timing-Peak-DRAM-Use-In-R-With-Perl-Part-2.html) only tracks the main processes memory? not a perl guy nor a RSS guy so may not be understanding it). For using `memprof`, I summed the memory for all processes at each time (R, bash, python, etc; whatever it logged) and used that as the total function call memory usage. The initial entry was used as baseline (after data has been loaded and before function called) and then the maximum RSS of all processes at a time point was used as the maximum with the difference being that due to the function call. 
+
+The values between the two trend in a similar direction for most except `BiocSingular` which is wildly different. Since I do not know which is 'correct' or if there is a real 'correct' way to measure memory, we have collectively removed the memory comparison from the manuscript. All the scripts and files can be found in the github repository at `docs_acs/paper` where the `*.R` are the scripts used to prep and run the profiling, `*.csv` are the outputs of the `memprof` package where the main process and spawned processes' memory are logged, and the `*.rds` are those from using the perl method. Also, removed mention of memory usage in the README to avoid confusion. 
+
+For `BiocSingular`, there were errors when running inside the `profmem::profmem` call so a custom `bench_time_mem()` was used here to skip memory tracking and only do time (can see in the 2.08 script). Further, because running on Mac I suspect?, had to modify the `monitor_memory.pl` script to use explicit "dash flag" instead of what is provided in your github.io otherwise would get errors saying that "p" isn't a valid option for `ps`. The updated `monitor_memory.pl` is also in `docs_acs/paper`. 
+
+> Point 5: If keeping memory allocation points, touch on relationship between memory allocation and time in manuscript and Readme of repository
+
+Have removed the memory stuff from the paper altogether and because of this, have changed Figure 1 to show the time taken for different algorithms (over 10 reps) as a boxplot rather than point plot of time vs memory.
+
+> Point 6: tutorial/article link and benchmarking script link for pointing no where
+
+This is because of the differences between the github.io and the regular github repository, and me trying to use dynamic links. I have updated the tutorial link from the repository Readme to have the full github.io link path. As for the benchmarking script (assuming within the manuscript), this is due to github inserting `blob/main/` into the path. We have updated the text to reflect this and the link should now be accessible.
+https://github.com/ACSoupir/FastPCA/blob/main/docs_acs/paper/benchmarking_script.R
+
+> Point 7: Stating "R-native" is misleading
+
+Thank you for mentioning this. We agree that "R-native" is not an appropriate description. Have rephrased it slightly to maintain the intention:
+"`FastPCA` provides an accelerated route to principal component analysis on large matrices by coupling randomized SVD with `torch`/`pytorch` backends and a familiar R API." As you state, everything has existed or exists in it's separate components but the novelty of this package is giving users *in R* the familiar interface to use the power of the highly optimized `libtorch` specifically for PCA analyses. 
+
[…]
