### 2026-01-29 #9706 @chrisarg
My comments were all addressed!
Thank you for putting the response together.

Memory management in software that call external library is always a
problem. I do quite a bit of Perl (also a garbage collected language) and
the only reliable way is to track the RSS (and even that is problematic if
the library links to another library).

Enjoy your weekend!
