### 2026-01-30 #9706 @ACSoupir
Yeah, I definitely have a better appreciation for the memory tracking and need to learn more! Thank you so much for the comments and suggestions. We definitely have benefited from them!

Hope you enjoy your weekend as well! February here we come!
