### 2026-02-09 #9706 @rich2355
Thanks @chrisarg, if there is no claim, this item also needs to be checked off.

> **Performance**: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

I have done this for you, so if no further questions, you officially concluded your review.

Thanks again for your time!
