### 2026-02-11 #9706 @dhjx1996
The paper and package look good to me! My remaining nitpick is that @ACSoupir should provide some kind of contact information; I still do not see Community Guidelines, or what looks to me like Community Guidelines, but I think most important is simply to give the community ways to reach out to you. Nonetheless, I have checked off the rest of my checklist and I conclude my review @rich2355
