### 2026-02-11 #9706 @ACSoupir
@dhjx1996 re-Community Guidelines: are you thinking in the README of the repository (also then landing page of of the github.io) are do you think there should be more in the paper under the Community Guidelines. Have added:

> Contributions by others is welcome for `FastPCA` as we believe it is a significant improvement to the field, allowing more people to analyze large datasets locally. To contribute, pull requests can be made to allow review before merging. Issues found should be submitted to https://github.com/ACSoupir/FastPCA/issues so we can fix them, improving user experience.

at the end of the manuscript but want to make sure I adequately address your concern/statement.

In the mean time, will add the above blurb to the repository README which will reflect on the github.io page as well.

EDIT:
I see. The markdown was updated but the paper itself (pdf, md) did not update. Have since re-built the paper available here:
https://github.com/ACSoupir/FastPCA/actions/runs/21910187236/artifacts/5467188224

And added community guidelines to the readme/github.io. https://acsoupir.github.io/FastPCA/index.html at the bottom this can be seen.
