+++ NEWS.md
@@ -1,5 +1,14 @@
 # FastPCA (development version)
 
+## Version 0.0.3
+
+* Added tests
+* Added updates to vignettes for "Suggests" packages
+* Removed memory benchmarks
+* Review comments 1
+  * https://github.com/openjournals/joss-reviews/issues/9706#issuecomment-3735025935
+  * https://github.com/openjournals/joss-reviews/issues/9706#issuecomment-3733686184
+
 ## Version 0.0.2
 
 * Added spatial transcriptomics tutorial
