### 2026-02-11 b8311cb
community guidelines paper
