+++ paper/paper.md
@@ -37,7 +37,7 @@ affiliations:
   - name: Department of Bioinformatics and Biostatistics, Moffitt Cancer Center
     index: 5
 citation_author: Ward et al
-date: "26 January, 2026"
+date: "11 February, 2026"
 year: "2026"
 bibliography: paper.bib
 output: 
