### 2026-02-11 #9706 @dhjx1996
@ACSoupir Ah, I did not realize you intended to put the Community Guidelines into the paper. No, I think placing it in your repository README is much better. In any case the most important thing is for people who want to contact you regarding `FastPCA` to know how to do so. I have no further comments or suggestions :)
