### 2026-02-11 d42f46d
remove community guidelines from paper
