### 2026-02-11 #9706 @ACSoupir
Ohhh that makes sense. Okay. I have removed it from the end of the manuscript and kept it in the README. Thank you for the clarification! Sorry for the goof.

- removed from manuscript: https://github.com/ACSoupir/FastPCA/actions/runs/21914240149/artifacts/5469017369
- have text in the [README](https://github.com/ACSoupir/FastPCA/blob/main/README.md) and [github.io](https://acsoupir.github.io/FastPCA/index.html) page.
