### 2026-02-11 #9706 @rich2355
Hi @ACSoupir, paper looks good to me! Before I proceed, could you please

- Check the four SKIP DOIs, if they have DOIs, please add them in the paper; if no, feel free to skip.

Thanks!
