+++ paper/paper.bib
@@ -270,14 +270,6 @@ abstract = {We present qrpca, a fast and scalable QR-decomposition principal com
   url = {https://doi.org/10.1093/bioinformatics/bty185},
 }
 
-@Manual{memprof,
-    title = {memprof: Profile memory over time of system or a process},
-    author = {Robert Ashton},
-    year = {2026},
-    note = {R package version 0.0.1, commit 1777ddfb3adbe408538f8fa424c9959bf39ea0ad},
-    url = {https://github.com/mrc-ide/memprof},
-  }
-
 @ARTICLE{Eckart1936-mg,
   title     = "The approximation of one matrix by another of lower rank",
   author    = "Eckart, Carl and Young, Gale",
@@ -299,17 +291,9 @@ abstract = {We present qrpca, a fast and scalable QR-decomposition principal com
   pages     = "211--218",
   month     =  sep,
   year      =  1936,
+  doi       = {10.1007/BF02288367},
   language  = "en"
 }
-
-@Manual{desctools,
-    title = {DescTools: Tools for Descriptive Statistics},
-    author = {Andri Signorell},
-    year = {2025},
-    note = {R package version 0.99.60},
-    url = {https://CRAN.R-project.org/package=DescTools},
-  }
-
 
 @ARTICLE{Akoglu2018-pq,
   title     = "User's guide to correlation coefficients",
@@ -333,6 +317,7 @@ abstract = {We present qrpca, a fast and scalable QR-decomposition principal com
   pages     = "91--93",
   month     =  sep,
   year      =  2018,
+  doi       = {10.1016/j.tjem.2018.08.001},
   keywords  = "Correlation coefficient; Cramer's; Interpretation; Lin's;
                Pearson's; Spearman's",
   copyright = "http://creativecommons.org/licenses/by-nc-nd/4.0/",
