### 2026-02-13 #9706 @ACSoupir
Hey, @rich2355!

For the DOIs:
1. `memprof` and `desctools` are no longer in the manuscript but is in the bib, have removed it.
2. have found and added DOIs for the other two now

Have also reran the check just above here.

Tag is [v0.9.0](https://github.com/ACSoupir/FastPCA/releases/tag/v0.9.0).

Have uploaded to Zenodo with the v0.9.0 release/tag with the DOI https://doi.org/10.5281/zenodo.18626773.
