### 2026-02-13 #9706 @rich2355
Hi @ACSoupir, thanks for your work! The DOIs look good to me. However, I think there are some inconsistencies elsewhere. Specifically,

1. The version information in Zenodo is v1, here you said v0.9.0.
2. The author list in Zenodo is not the same as that in your paper.

Could you please double check on these (and other possible inconsistencies)? Thanks again!
