### 2026-02-13 #9706 @ACSoupir
Hey, @rich2355. Thanks for your response.

Didn't know could change the versions to custom values for Zenodo. The more you know!
I have also added everyone for the paper to the Zenodo page.
Updated DOI: https://doi.org/10.5281/zenodo.18626988

Let me know if there's other things you'd like from me/us. Really appreciate your help.
