### 2026-02-13 #9706 @rich2355
Hi @ACSoupir, I will defer to the AEiC for further actions. But for now, I would like to say congrats to you and also many thanks to the fantastic two reviewers @chrisarg, @dhjx1996 (and if I had found four reviewers, you would be the Fantastic Four :) ). I can feel the warmness and kindness, and your communication went very well. This made my editorial work enjoyable and substantially reduced my workload.
