### 2026-02-13 #9706 @ACSoupir
Thank you, @rich2355! I really like the open reviews that allow a back and forth dialog.

And super appreciate the recommendations and comments from @chrisarg and @dhjx1996. We think it's made the work and package stronger, hopefully useful to the community!

If there's anything more that we can help with, feel free to ping me. We will try and tackle it ASAP!
