### 2026-02-13 #9706 @ACSoupir
@rich2355, looking at the proof, Brooke wanted to add her professor appointment affiliation. What would be the easiest way to do this?
