### 2026-02-13 #9706 @rich2355
@ACSoupir Please edit the `paper.md` file. The AEiC will recompile the whole submission so that the change will be in effect.
