### 2026-02-13 84cc257
Update affiliations for authors in paper.md
