### 2026-03-06 #9706 @crvernon
@ACSoupir thank you for your patience while I have been getting caught back up. I just need you to address the following before I accept this for publication:

In your archive:

- [x] Please edit the metadata of your Zenodo archive to order the authors exactly the way you have them listed in the paper.

- [x] Please remove the backticks from the name of your package in the title of your paper.  These will cause some formatting issues when we publish.
- [x]   LINE 12:  Your in-text citation "...SVD, Halko, Martinsson, & Tropp (2011))..." when used in this context should appear as "...SVD, (Halko, Martinsson, & Tropp, 2011)..." You can modify this in your paper by using brackets around your citation like:  `...decomposition (SVD, [@Halko-ILBRA]) with...` instead of what you have now which is `...decomposition (SVD, @Halko-ILBRA) with...`.
- [x] LINE 20:  Same in-text citation issue here where "...biological data (Lim, Wang, Buzdin, & Li (2025))." should appear as "...biological data (Lim, Wang, Buzdin, & Li (2025)."  You will not need to manually add in parenthesis in this case if you wrap the citation in brackets in your paper.md.
- [x] I see the citation issue repeated throughout the paper.  Please address each of these.  The general rule on the format style you choose would be the following:  if you are directly discussing a researcher's work, use (fake example):  "We found that the methods presented in Doe & Jones (2025) gave us the ability to..." where when indirectly using a work or works to support your statement you would use:  "There are many methods that have been presented on topic X (Doe & Jones, 2025; Smith et al., 2025)."
- [x] LINE 28:  The "p" in "python" should be capitalized.
- [x] LINE 42:  The "p" in "python" should be capitalized.
- [x] LINE 45:  The "p" in "python" should be capitalized.
- [x] LINES 73-73:  When clicked, the link produces a 404.
- [x] LINE 89:  You state, "...time to 2.44 (2.44 to 2.44) seconds." Is what you have in parenthesis correct?  Same for LINE 90-91.  There is no between for 2.44 and 2.44.  I would restate this if 2.44 was the singular value.
- [x] LINE 171:  The "p" in "python" should be capitalized.  You can do this in your bib file by placing characters you wish to maintain the formatting of in curly brackets.

Let me know when these changes have been made, and after I confirm, I will accept.  Thank you.
