+++ paper/paper.Rmd
@@ -18,13 +18,13 @@ authors:
     affiliation: 3
   - name: Brooke L Fridley
     orcid: 0000-0001-7739-7956
-    affiliation: 4
+    affiliation: "4, 5"
   - name: Steven Eschrich
     orcid: 0000-0002-9833-2788
-    affiliation: 5
+    affiliation: 6
   - name: Alex C Soupir
     orcid: 0000-0003-1251-9179
-    affiliation: "2, 5" # (Multiple affiliations must be quoted)
+    affiliation: "2, 6" # (Multiple affiliations must be quoted)
 affiliations:
   - name: Department of Cutaneous Oncology, Moffitt Cancer Center
     index: 1
@@ -34,8 +34,10 @@ affiliations:
     index: 3
   - name: Health Services & Outcomes Research, Children's Mercy
     index: 4
-  - name: Department of Bioinformatics and Biostatistics, Moffitt Cancer Center
+  - name: Department of Pediatrics, University of Missouri-Kansas City School of Medicine
     index: 5
+  - name: Department of Bioinformatics and Biostatistics, Moffitt Cancer Center
+    index: 6
 citation_author: Ward et al
 date: "`r format(Sys.time(), '%d %B, %Y')`"
 year: "`r format(Sys.time(), '%Y')`"
@@ -49,19 +51,19 @@ journal: JOSS
 
 # Summary
 
-The `FastPCA` package provides an interface to optimized matrix multiplication libraries (`libtorch`) for the purpose of singular value decomposition. Using `FastPCA` to perform randomized singular value decomposition (SVD, @Halko-ILBRA) with the `torch` or `pytorch` backend drastically reduces the computational time compared to base R `prcomp` and other truncated singular value decomposition. Developed for biological data such as single-cell RNA-sequencing, spatial transcriptomics, or matrix-assisted laser desorption/ionization (MALDI imaging), `FastPCA` can efficiently and accurately identify the leading singular values in high-dimensional data.
+The `FastPCA` package provides an interface to optimized matrix multiplication libraries (`libtorch`) for the purpose of singular value decomposition. Using `FastPCA` to perform randomized singular value decomposition (SVD, [@Halko-ILBRA]) with the `torch` or `pytorch` backend drastically reduces the computational time compared to base R `prcomp` and other truncated singular value decomposition. Developed for biological data such as single-cell RNA-sequencing, spatial transcriptomics, or matrix-assisted laser desorption/ionization (MALDI imaging), `FastPCA` can efficiently and accurately identify the leading singular values in high-dimensional data.
 
 # Statement of Need
 
-Over the last decade, advances in single-cell and spatial profiling technologies have dramatically increased the scale and resolution of biological data (@Lim2025-vb). Multiple custom and commercially available technologies produce tens of thousands of samples (e.g., spots, cells, pixels) and thousands of measured features (i.e., genes, peaks, etc) per sample which can require tens to hundreds of gigabytes (@Zaima2010-gy). Creating linear or non-linear combinations of the input space in lower-dimensional space enables more efficient downstream analysis. However, this is typically done with packages like `irlba` (@Baglama2011-ob) where the process of identifying singular values is iterative and single-thread limited. Other packages can improve efficiency by using an API for R (such as `reticulate` for python) but is not multithreaded without modification to the source code (@Li2023-aa). The `qrpca` R package (@De_Souza2022-lk) uses `torch` (@Falbel2020-wq; @Paszke2019-go) to perform QR-based principal component analysis (PCA), but doesn't produce truncated singular values resulting in incredibly large memory requirements for large matrices. For biological studies,  parameter tuning and rapid iteration (e.g., normalization methods, subclustering for identifying specific cell types) are essential to identify biologically meaningful signals, which typically reside in early dimensions, enabling the avoidance of the full decomposition.
[…]
