+++ paper/paper.bib
@@ -225,7 +225,7 @@ abstract = {We present qrpca, a fast and scalable QR-decomposition principal com
 }
 
 @MISC{Ushey2017-mx,
-  title     = "reticulate: Interface to 'Python'",
+  title     = "reticulate: Interface to '{P}ython'",
   author    = "Ushey, Kevin and Allaire, J J and Tang, Yuan",
   publisher = "The R Foundation",
   month     =  mar,
