### 2026-03-06 #9706 @ACSoupir
Thanks, @crvernon . Didn't know the bracket thing so glad to have learned that.

Have also removed the back ticks from the title, and edited the zenodo archive to match author order for the paper.

Fixed the "python" to "Python"

The script was moved during revision if i remember correctly. Link now is correct.

Have changed both 2.44 cases to remove the phrasing of the other times and just state the single time.

Cool trick with the curly brackets again! Thank you.

Let me know if you'd like anything else.
https://github.com/ACSoupir/FastPCA/actions/runs/22777304053/artifacts/5803432733
