### 2026-03-06 #9706 @crvernon
@ACSoupir the backticks are still in your title around the package name:  see https://raw.githubusercontent.com/ACSoupir/FastPCA/refs/heads/main/paper/paper.md
