### 2026-03-06 #9706 @ACSoupir
@crvernon have updated:
https://raw.githubusercontent.com/ACSoupir/FastPCA/refs/heads/main/paper/paper.md
